Admin has more options on which clips to allow and deny
The admin can now choose to only allow clips from specific channel,
clips that were made during a specific date or later, whether or not
the user who is uploading the clip also needs to be the one who has
clipped, and also added a cutoff point for being able to upload clips.

Fixed some timezone issues caused in the Clips model.