Uploaded user now visible during ranking and result viewing
When ranking clips and viewing the results the twitch username
of the person who sent the clip is now visible. This name is the
username saved in the database and does not make an API call
to twitch servers using the user id. As a result, the twitch
username may be outdated.